+++ common/src/main/java/io/github/mortuusars/horseman/mixin/render/LivingEntityRendererMixin.java
@@ -1,15 +1,11 @@
 package io.github.mortuusars.horseman.mixin.render;
 
 import com.llamalad7.mixinextras.sugar.Local;
-import com.llamalad7.mixinextras.sugar.Share;
-import com.llamalad7.mixinextras.sugar.ref.LocalDoubleRef;
-import com.llamalad7.mixinextras.sugar.ref.LocalIntRef;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import io.github.mortuusars.horseman.client.HorseRenderUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.EntityModel;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
@@ -23,7 +19,6 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(value = LivingEntityRenderer.class, priority = 950)
@@ -32,18 +27,11 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, M extend
         super(context);
     }
 
-    @Inject(method = "render(Lnet/minecraft/world/entity/LivingEntity;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V",
-            at = @At("HEAD"))
-    void setAlpha(T entity, float entityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int packedLight, CallbackInfo ci, @Share("alpha") LocalDoubleRef alpha) {
-        if (entity instanceof AbstractHorse) {
-            alpha.set(HorseRenderUtils.getAlpha(entity));
-        }
-    }
-
     @Redirect(method = "render(Lnet/minecraft/world/entity/LivingEntity;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/client/model/EntityModel;renderToBuffer(Lcom/mojang/blaze3d/vertex/PoseStack;Lcom/mojang/blaze3d/vertex/VertexConsumer;IIFFFF)V"))
     private void onRender(EntityModel<?> instance, PoseStack poseStack, VertexConsumer vertexConsumer,
-                          int packedLight, int packedOverlay, float r, float g, float b, float a, @Local(argsOnly = true) T entity, @Share("alpha") LocalDoubleRef alpha) {
+                          int packedLight, int packedOverlay, float r, float g, float b, float a,
+                          @Local(argsOnly = true) T entity) {
         if (entity instanceof AbstractHorse) {
             if (HorseRenderUtils.isJeb(entity)) {
                 int index = entity.tickCount / 25 + entity.getId();
@@ -64,7 +52,10 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, M extend
                 b = Mth.clamp(b * 2, 0, 1);
             }
 
-            a = (float) Mth.clamp(a * alpha.get(), 0.0, 1.0);
+            double alphaVal = HorseRenderUtils.getAlpha(entity);
+            if (alphaVal != 0.0) {
+                a = (float) Mth.clamp(a * (alphaVal - 0.1), 0.0, 1.0);
+            }
         }
 
         instance.renderToBuffer(poseStack, vertexConsumer, packedLight, packedOverlay, r, g, b, a);
@@ -72,8 +63,8 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, M extend
 
     @Inject(method = "getRenderType", at = @At("HEAD"), cancellable = true)
     private void getRenderType(T entity, boolean bodyVisible, boolean translucent, boolean glowing,
-                               CallbackInfoReturnable<RenderType> cir, @Share("alpha") LocalDoubleRef alphaRef) {
-        if (entity instanceof AbstractHorse && alphaRef.get() != 1.0) {
+                               CallbackInfoReturnable<RenderType> cir) {
+        if (entity instanceof AbstractHorse && HorseRenderUtils.getAlpha(entity) != 1.0) {
             cir.setReturnValue(RenderType.entityTranslucent(getTextureLocation(entity)));
         }
     }