+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.1 - 2025-04-18
+- [Forge] - Made new features work with Realistic Horse Genetics (fading, creative taming)
+  - Fixed Genetic Horses always rendering slightly translucent.
+
 ## 1.3.0 - 2025-04-18
 - Improved mount gui - shows player's hunger bar, xp bar and xp levels. Horse jump bar will only show when jumping.
 - When riding a horse - jump meter will no longer fill up when mount is in water.